Create 03-cone.js
Cone
Write a JS function to calculate a cone’s volume and total surface area by given height and radius at the base.
The input comes as two number arguments. The first element is the cone’s radius and the second is its height.
The output should be printed to the console on a new line for every result. The result should be formatted to the fourth decimal point
Examples
 
Input	Output		Input	Output	   
3,
5	volume = 47.1239
area = 83.2298		3.3,
7.8	volume = 88.9511
area = 122.0159	 
Hints