Create 08-validity-checker.js
Validity Checker
Write a program that receives two points in the format [x1, y1, x2, y2]. Check if the distance between each point and the start of the cartesian coordinate system (0, 0) is valid. A distance between two points is considered valid, if it is an integer value. 
In case a distance is valid, print"{x1, y1} to {x2, y2} is valid"
If the distance is invalid, print "{x1, y1} to {x2, y2} is invalid"
The order of comparisons should always be first {x1, y1} to {0, 0}, then {x2, y2} to {0, 0} and finally {x1, y1} to {x2, y2}. 
The input consists of two points given as an array of numbers.
For each comparison print either "{x1, y1} to {x2, y2} is valid" if the distance is valid, or "{x1, y1} to {x2, y2} is invalid" if it is invalid.
Examples
Input	Output
[3, 0, 0, 4]	{3, 0} to {0, 0} is valid
{0, 4} to {0, 0} is valid
{3, 0} to {0, 4} is valid
[2, 1, 1, 1]	{2, 1} to {0, 0} is invalid
{1, 1} to {0, 0} is invalid
{2, 1} to {1, 1} is valid