Create 01-print-array-with-a-given-delimeter.js
Print an Array with a Given Delimiter
The input comes as an array of strings. The last element of the array is the delimiter.
The output is the same array, printed on the console, each element separated from the others by the given delimiter.
Examples
Input	Output		Input	Output
['One', 
'Two', 
'Three', 
'Four', 
'Five', 
'-']	One-Two-Three-Four-Five		['How about no?', 
'I',
'will', 
'not', 
'do', 
'it!', 
'_']	How about no?_I_will_not_do_it!