Create 02-print-evry-n-element-from-arr.js
Print Every N-th Element from an Array 
The input comes as an array of strings. The last element is N - the step.
The output is every element on the N-th step starting from the first one. If the step is "3", you need to print the 1-st, the 4-th, the 7-th … and so on, until you reach the end of the array. The elements must be printed each on a new line.
Example
Input	Output		Input	Output		Input	Output
['5', 
'20', 
'31', 
'4', 
'20', 
'2']	5
31
20		['dsa',
'asd', 
'test', 
'tset', 
'2']
	dsa
test	 	['1', 
'2',
'3', 
'4', 
'5', 
'6']	1
Hints
Use what you’ve seen from the previous problem to extract the last element of the array.
Create a step variable to hold the given step of the array. Then print all the elements with a for loop, incrementing the loop variable with the value of the step variable.