Create 01-function-validator.js
Error Handling
⦁	Request Validator
Write a function that validates an HTTP request object. The object has the properties method, uri, version and message. Your function will receive the object as a parameter and has to verify that each property meets the following requirements:
⦁	method - can be GET, POST, DELETE or CONNECT
⦁	uri - must be a valid resource address or an asterisk (*); a resource address is a combination of alphanumeric characters and periods; all letters are Latin; the URI cannot be an empty string
⦁	version - can be HTTP/0.9, HTTP/1.0, HTTP/1.1 or HTTP/2.0 supplied as a string
⦁	message - may contain any number or non-special characters;special characters are <, >, \, &, ', "
If a request is valid, return it unchanged. 
If any part fails the check, throw an Error with message "Invalid request header: Invalid {Method/URI/Version/Message}". 
Replace the part in curly braces with the relevant word. Note that some of the properties may be missing, in which case the request is invalid. Check the properties in the order in which they are listed above. If more than one property is invalid, throw an error for the first encountered.
Input / Output
Your function will receive an object as a parameter. Return the same object or throw an Error as described above as an output.
Examples
Input	Output
{
  method: 'GET',
  uri: 'svn.public.catalog',
  version: 'HTTP/1.1',
  message: ''
}	{
  method: 'GET',
  uri: 'svn.public.catalog',
  version: 'HTTP/1.1',
  message: ''
}
{
  method: 'OPTIONS',
  uri: 'git.master',
  version: 'HTTP/1.1',
  message: '-recursive'
}	Invalid request header: Invalid Method

{
  method: 'POST',
  uri: 'home.bash',
  message: 'rm -rf /*'
}	Invalid request header: Invalid Version
Hints
Since validating some of the fields may require the use of RegExp, you can check your expressions using the following samples:
URI
Valid	Invalid
svn.public.catalog
git.master
version1.0
for..of
.babelrc
c	%appdata%
apt-get
 
home$
define apps
"documents"
⦁	Note that the URI cannot be an empty string.
Message
Valid	Invalid
-recursive
rm -rf /*
hello world
https://svn.myservice.com/downloads/
%root%	<script>alert("xss vulnerable")</script>
\r\n
&copy;
"value"
'; DROP TABLE
⦁	Note that the message may be an empty string, but the property must still be present.