Create 02-Expedition-problems-66-100.js
Expedition Problems 
You will receive a text (string), which can contain all the ASCII symbols, including new lines and tabs. The location of the lost person and his message must be retrieved from this string. The text contains a keyword that indicates the beginning and the end of the message. The geographical coordinates come as a pair of longitude ("east") and latitude ("north") and each coordinate should meet the following conditions:
1.	It should start with "north"/"east", case-insensitive;
2.	The next symbols should be 2 digits, which form the whole part of the degrees;
3.	The whole part of the degrees is separated from the decimal part by "," and there may be any characters between them, except ","
4.	The decimal part consists of 6 digits
In case there is more than one longitude or latitude in the text, take the latter. The message is surrounded by the keyword, which will be the first argument from the input. The second argument from the input will be the text, containing both the location and the message of the lost person. See the examples below to understand how it works.
Input
The first argument contains the keyword and the second argument contains the text. There will always be at least one pair of coordinates.
Output
Print the latitude in a paragraph in the span with id "result" in the following format:
<degrees>.<decimal part> N
Print the second in a paragraph in the span with id "result" in the following format: 
<degrees>.<decimal part> E
On the last line, print the message in the following format:
Message: <message>
Example
Input
4ds
eaSt 19,432567noRt north east 53,123456north 43,3213454dsNot all those who wander are lost.4dsnorth 47,874532
Output
47.874532 N
19.432567 E
Message: Not all those who wander are lost.


Input
<>
o u%&lu43t&^ftgv><nortH4276hrv756dcc,  jytbu64574655k <>ThE sanDwich is iN the refrIGErator<>yl i75evEAsTer23,lfwe 987324tlblu6b
Output
42.645746 N
23.987324 E
Message: ThE sanDwich is iN the refrigerator
 Hints